+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 0.0.2 (TBA)
+* New feature: Support trade orderbooks
 * Improvement: Added unit tests
 * Improvement: Added subscribe / unsubsbscribe methods for candles and ticker to the ticketmanager
 * Bugfix: The ticks now a volume of 0